backend/requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

# Authentication & Security
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
email-validator>=2.1.0

# Face Recognition
deepface>=0.0.79
opencv-python>=4.8.0
tensorflow>=2.14.0
retina-face>=0.0.13

# LangChain & RAG
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langgraph>=0.0.20
langchain-google-genai>=1.0.0
google-generativeai>=0.3.0

# Vector Store & Search
faiss-cpu>=1.7.4
chromadb>=0.4.0
rank-bm25>=0.2.2

# Web Search & APIs
tavily-python>=0.3.0
httpx>=0.25.0
requests>=2.31.0

# ML для финансов
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
joblib>=1.3.0

# WebSocket
websockets>=12.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utils
pillow>=10.0.0

# Document Inspector (YOLOv8 for object detection)
ultralytics>=8.0.0
pymupdf>=1.23.0

